🎉 add : 이미지 다운로드 기능을 위한 유틸리티 함수 추가 및 기존 코드 리팩토링 #121
